refactor: 탭 UI 분리
